requirements.txt
hydra-core
PyYAML
python-dotenv
pyrootutils
opencv-python
numpy
pandas
transformers
fastapi
pydantic[email]
uvicorn
beanie
asyncio
colorlog
gunicorn
SQLAlchemy
sqlmodel
passlib
python-jose
python-multipart
